new dashboard analytics

+++ SADWebApi/Contracts/Sales/SalesSummaryDto.cs
@@ -1,7 +1,7 @@
 namespace Sad.Api.Contracts.Sales;
 
 public record SalesSummaryDto(
-  int Total,
-  int ThisMonth,
-  int Today
+  decimal Total,
+  decimal ThisMonth,
+  decimal Today
 );
